% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsf.R
\name{bsf}
\alias{bsf}
\title{Bootstrap function (bsf) for H+ Estimate}
\usage{
bsf(d, l, s, q)
}
\arguments{
\item{d}{(data, nxm)}

\item{l}{(table, pre-generated by hpb())}

\item{r}{(numeric, 1x1)}
}
\value{
One iteration of H+ bootstrap estimate
}
\description{
Estimates H+ for a random subset of the supplied data
to be replicated within the main hpb() function
Samples the observations (rows) of d and
the entries of l to generate a single bootstrap estimate.
}
\keyword{internal}
