% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_estm.R
\name{hp_estm}
\alias{hp_estm}
\title{Estimate H+}
\usage{
hp_estm(dis, lab, p = 0.005)
}
\arguments{
\item{dis}{dissimilarity (distance or matrix with dim nxn)}

\item{lab}{identity label (vector with length n)}

\item{p}{percentage of within- and between-cluster distances to sample (numeric)}
}
\value{
Estimated H+ given dis, lab, and p
}
\description{
Estimates H+ from a dissimilarity matrix (dis) and identity/label vector (lab).
Utilizes order statistics to algorithmically approximate H+ by sampling
a percentage (p) of within- and between-cluster distances.
}
\examples{
a <- sapply(1:100, function(i) rnorm(n=50,mean=1.0,sd=1))
l <- c(rep(0,50),rep(1,50))
d <- dist(t(a))
h <- hp_estm(dis=d,lab=l)
}
