% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_ests.R
\name{hp_ests}
\alias{hp_ests}
\title{Estimate H+}
\usage{
hp_ests(a, b, p = 0.005)
}
\arguments{
\item{a}{(vector, any length)}

\item{b}{(vector, any length)}

\item{p}{percentage of within- and between-cluster distances to sample (numeric)}
}
\value{
Estimated H+ given a, lab, and p
}
\description{
Estimates H+ for two sets (vectors) a and b.
Utilizes order statistics to algorithmically approximate H+ by sampling
a percentage (p) of within- and between-cluster distances.
}
\examples{
a <- sapply(1:100, function(i) rnorm(n=50,mean=1.0,sd=1))
b <- sapply(1:100, function(i) rnorm(n=50,mean=-1.0,sd=1))
h <- hp_ests(a=a,b=b)
}
