% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe.R
\name{hpe}
\alias{hpe}
\title{H+ discordance metric}
\usage{
hpe(A, B, D, L, p = 101, alg = "brute_force")
}
\arguments{
\item{A}{numeric vector containing a set of length n}

\item{B}{numeric vector containing a set of length n}

\item{D}{distance matrix of dimension nxn}

\item{L}{numeric or character vector of length n}

\item{p}{integer representing the number of percentiles}

\item{alg}{character string ("brute_force" or "grid_search") representing the choice of algorithm used to estimate H+}
}
\value{
list, h is the estimated H+ value.

gamma1 and gamma2 are plausible ranges for what \% of A (or Dw)

are strictly greater than B (or Db)
}
\description{
Estimates the H+ discordance metric
for either (1) two sets (vectors) A and B, or
(2) a dissimilarity matrix D and a label vector L.
Approximation is calculated using p+1 percentiles,
with an accuracy bound of 1/p.
}
\examples{
a <- rnorm(n=500, mean=0)
b <- rnorm(n=500 ,mean=1)
h <- hpe(A=a, B=b, p=101, alg="brute_force")

a <- sapply(1:500, function(i) rnorm(n=50, mean=0))
b <- sapply(1:500, function(i) rnorm(n=50, mean=0))
x <- cbind(a,b)
d <- dist(t(x))
l <- c(rep(0,500), rep(1,500))
h <- hpe(D=d, L=l, p=101, alg="brute_force")

}
