% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg1.R
\name{alg1}
\alias{alg1}
\title{Algorithm 1 ("brute_force") for H+ Estimate}
\usage{
alg1(qA, qB, p)
}
\arguments{
\item{qA}{(vector, px1)}

\item{qB}{(vector, px1)}
}
\value{
Estimated H+, un-standardized (numeric vector)
}
\description{
Estimates H+ for two sets (vectors) A and B, or
a dissimilarity matrix D and a label vector L.
Approximation is calculated using p+1 percentiles,
with an accuracy bound of 1/p.
}
\keyword{internal}
