% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe.R
\name{hpe}
\alias{hpe}
\title{H+ Estimate}
\usage{
hpe(A, B, D, L, p = 101)
}
\arguments{
\item{A}{(numeric vector)}

\item{B}{(numeric vector)}

\item{D}{(dist or symmetric matrix, nxn)}

\item{L}{(numeric or character vector, nx1)}

\item{p}{(integer)}
}
\value{
Estimated H+ (numeric)
}
\description{
Estimates H+ for two sets (vectors) A and B, or
a dissimilarity matrix D and a label vector L.
Approximation is calculated using p+1 percentiles,
with an accuracy bound of 1/p.
}
\examples{
a <- sapply(1:100, function(i) rnorm(n=50,mean=1.0,sd=1))
b <- sapply(1:100, function(i) rnorm(n=50,mean=-1.0,sd=1))
h <- hpe(a=a,b=b,p=101)
a <- cbind(a,b)
d <- dist(t(a))
l <- c(rep(0,100),rep(1,100))
h <- hpe(D=d,l=b)
}
