% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe.R
\name{hpe}
\alias{hpe}
\title{H+ Estimate}
\usage{
hpe(A, B, D, L, p = 101, alg = 1)
}
\arguments{
\item{A}{(numeric vector)}

\item{B}{(numeric vector)}

\item{D}{(dist, nxn)}

\item{L}{(numeric or character vector, nx1)}

\item{p}{(integer)}

\item{alg}{(integer), Algorithm 1 or 2}
}
\value{
Estimated H+ (numeric)
}
\description{
Estimates H+ for two sets (vectors) A and B, or
a dissimilarity matrix D and a label vector L.
Approximation is calculated using p+1 percentiles,
with an accuracy bound of 1/p.
}
\examples{
a <- rnorm(n=500,mean=0)
b <- rnorm(n=500,mean=1
h <- hpe(a=a,b=b,p=101,alg=1)

a <- sapply(1:500, function(i) rnorm(n=50,mean=0))
b <- sapply(1:500, function(i) rnorm(n=50,mean=0))
x <- cbind(a,b)
d <- dist(t(x))
l <- c(rep(0,500),rep(1,500))
h <- hpe(D=d,l=b,p=101,alg=1)
}
